Introduce a test case for shallow copy
Introduce a test case to check a shallow copy in SBFL Formular - ER1a
I don't have know whether it is intended or not to use a shallow copy in ER1a
but I believe that it could be a kind of mistake.
Before modifying the algorithm to use a `copy` method, simple test case is prepared to check current functionalities.